Moved timer to CoreWorkflow
