+++ connect/workflows/core.py
@@ -3,16 +3,19 @@ core.py
 
 Provides the base LinuxForHealth workflow definition.
 """
+import inspect
 import json
+import functools
 import logging
 import connect.clients.nats as nats
+import time
 import uuid
 import xworkflows
 from datetime import datetime
 from fastapi import Response
 from httpx import AsyncClient
 from connect.clients.kafka import get_kafka_producer, KafkaCallback
-from connect.config import nats_sync_subject
+from connect.config import get_settings, nats_sync_subject
 from connect.exceptions import LFHError
 from connect.routes.data import LinuxForHealthDataRecordResponse
 from connect.support.encoding import (
@@ -21,7 +24,6 @@ from connect.support.encoding import (
     decode_to_str,
     ConnectEncoder,
 )
-from connect.support.timer import timer
 
 
 logger = logging.getLogger(__name__)
@@ -74,8 +76,33 @@ class CoreWorkflow(xworkflows.WorkflowEnabled):
         self.lfh_id = kwargs["lfh_id"]
         self.transmit_server = kwargs.get("transmit_server", None)
         self.do_sync = kwargs.get("do_sync", True)
+        self.uuid = str(uuid.uuid4())
 
     state = CoreWorkflowDef()
+
+    def timer(func):
+        """Decorator to print the elapsed runtime of the decorated function"""
+
+        @functools.wraps(func)
+        async def timer_wrapper(*args, **kwargs):
+            settings = get_settings()
+            if settings.connect_timing_enabled:
+                start_time = time.time()
+
+            if inspect.iscoroutinefunction(func):
+                result = await func(*args, **kwargs)
+            else:
+                result = func(*args, **kwargs)
+
+            if settings.connect_timing_enabled:
+                run_time = time.time() - start_time
+                self = args[0]
+                logger.trace(
+                    f"{func.__name__}() id = {self.uuid} elapsed time = {run_time:.7f}s"
+                )
+            return result
+
+        return timer_wrapper
 
     @xworkflows.transition("do_validate")
     @timer
@@ -124,7 +151,7 @@ class CoreWorkflow(xworkflows.WorkflowEnabled):
             encoded_data = encode_from_str(self.message)
 
         message = {
-            "uuid": str(uuid.uuid4()),
+            "uuid": self.uuid,
             "lfh_id": self.lfh_id,
             "creation_date": str(datetime.utcnow().replace(microsecond=0)) + "Z",
             "store_date": str(datetime.utcnow().replace(microsecond=0)) + "Z",
